♻️ minor fixes
