+++ src/main.py
@@ -23,15 +23,19 @@ def init_api() -> FastAPI:
         title="Que Account",
         version="0.1.0",
     )
+
+    return app
+
+
+def init_services(app: FastAPI) -> None:
     setup_middlewares(app)
     configure_logging()
     container = Container()
     app.container = container
     setup_routes(app)
-    return app
 
 
-async def run_server(app: FastAPI) -> None:
+async def start_server(app: FastAPI) -> None:
     config = uvicorn.Config(
         app,
         host="127.0.0.1",
@@ -46,7 +50,8 @@ async def run_server(app: FastAPI) -> None:
 
 async def main() -> None:
     app = init_api()
-    await run_server(app)
+    init_services(app)
+    await start_server(app)
 
 
 if __name__ == "__main__":
